requirements.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2020.12.5
charset-normalizer==2.0.4
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.2
django-bower==5.2.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==7.1
django-crispy-forms==1.10.0
django-dbbackup==3.3.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
# django-import-export==2.5.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.1.1
django-simple-history==3.0.0
django-webcam==0.4.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.0.4
icalendar==4.0.7
idna==3.2
isort==5.8.0
jsmin==2.2.2
jsonfield==3.1.0
lazy-object-proxy==1.6.0
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.0.1
psycopg2-binary==2.8.6
pylint==2.8.2
python-barcode==0.13.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
qrcode==6.1
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
tablib==3.0.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.6
uuid==1.30
# whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0